backend/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
pymysql
cryptography
python-jose
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
redis
requests
aiohttp
pydantic>=2.0.0